refactor(fields): migrate CurrencyField to the shared currency resolver
Phase 2b migration (batch 1). `CurrencyField` (the currency form input +
readonly display) read ONLY `field.currency`, so a field declaring its currency
via `currencyConfig.defaultCurrency`, or relying on the tenant default, rendered
symbol-less.

- Extract `resolveFieldCurrency` from the giant `index.tsx` into a leaf
  `fields/src/currency.ts` (re-exported) so widget components can import it
  without an `index ↔ widget` cycle. Better hygiene; no API change.
- `CurrencyField` now resolves via `resolveFieldCurrency(field, useLocalization()
  .currency)` — field currency → currencyConfig → tenant default → plain number.

Tests: resolveFieldCurrency 4 + standard-widgets 18 pass; type-check green.
Behavior unchanged when no provider / explicit field currency (the common case).
